refactor: 경로 및 요금 테스트 파일 분리
